Rename: 重命名 three_axis_charts 包
